+++ packages/chaingraph-nodes/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @badaitech/chaingraph-nodes
 
+## 0.1.17
+
+### Patch Changes
+
+- Add port documentation and debug value tooltip. Make port order configurable
+- Updated dependencies
+  - @badaitech/chaingraph-types@0.1.17
+
 ## 0.1.16
 
 ### Patch Changes